# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Encois Gateway API
  version: 1.0.0
paths:
...
  /api/v1/workflows:
    post:
      operationId: startWorkflow
      summary: Start a registered workflow or generic Blueprint execution
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowStartRequest'
      responses:
        '202':
          description: Workflow accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionResponse'
    get:
      operationId: listWorkflows
      summary: List workflow projections visible in the caller's effective scope
      security:
        - bearerAuth: []
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 160 } }
        - { name: status, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [updated-desc, updated-asc, name-asc, status] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Scoped workflow projections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionListResponse'
...
  /api/v1/internal/coordinator/plans/validate:
    post:
      operationId: validateCoordinatorWorkflowChangePlan
      summary: Validate a Coordinator proposal through the private service boundary
      security:
        - serviceTokenAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowChangePlanRequest'
      responses:
        '200':
          description: Plan validated but not applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPlanValidationResponse'
  /api/v1/internal/coordinator/plans:
    post:
      operationId: submitCoordinatorWorkflowChangePlan
      summary: Submit a Coordinator proposal through the private service boundary
      security:
        - serviceTokenAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowChangePlanRequest'
      responses:
        '202':
          description: Proposal persisted and awaiting approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPlanResponse'
  /api/v1/internal/coordinator/workflows:
    post:
      operationId: startCoordinatorApprovedWorkflow
      summary: Start a generic Workflow from an approved Blueprint registry snapshot
      security:
        - serviceTokenAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowStartRequest'
      responses:
        '202':
          description: Approved workflow accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionResponse'
...
  /api/v1/workflows/plans/validate:
    post:
      operationId: validateWorkflowChangePlan
      summary: Validate a proposed workflow change without applying it
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowChangePlanRequest'
      responses:
        '200':
          description: Plan validated but not applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPlanValidationResponse'
  /api/v1/workflows/plans:
    get:
      operationId: listWorkflowChangePlans
      summary: List tenant-visible persisted workflow change proposals
      security:
        - bearerAuth: []
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 160 } }
        - { name: status, in: query, schema: { type: string, enum: [all, proposed, approved, rejected, applied, expired] } }
        - { name: sort, in: query, schema: { type: string, enum: [updated-desc, updated-asc, name-asc, status] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Persisted proposals visible in the caller's organization and execution scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPlanListResponse'
    post:
      operationId: submitWorkflowChangePlan
      summary: Persist a validated workflow change proposal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowChangePlanRequest'
      responses:
        '202':
          description: New proposal persisted and awaiting approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPlanResponse'
        '200':
          description: Idempotent replay of an existing proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPlanResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    serviceTokenAuth:
      type: apiKey
      in: header
      name: X-Encois-Service-Token
  parameters:
    OrganizationHeader:
      name: X-Organization-ID
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    ExecutionScope:
      type: object
      additionalProperties: false
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          items: { type: string }
...
    WorkflowStartRequest:
      type: object
      additionalProperties: false
      required: [workflowType]
      properties:
        workflowType:
          type: string
          enum: [encois.dynamic.v1, CoordinatorWorkflow, BootstrapProjectWorkflow]
        version: { type: string, maxLength: 128 }
        key: { type: string, maxLength: 128 }
        blueprintId: { type: string, maxLength: 128 }
        blueprintVersion: { type: string, maxLength: 128 }
        idempotencyKey: { type: string, maxLength: 128 }
        input: { type: object }
        scope: { type: object }
        blueprint: { type: object }
    WorkflowExecutionResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [workflowId, workflowType, namespace, taskQueue, status, organizationId, createdAt, updatedAt]
          properties:
            workflowId: { type: string }
            runId: { type: string }
            workflowType: { type: string }
            namespace: { type: string }
            taskQueue: { type: string }
            blueprintId: { type: string }
            status: { type: string, enum: [queued, running, waiting, paused, partial, failed, completed, cancelled] }
            statusReason:
              type: string
              enum: [temporary_error, missing_credentials, human_approval, capability_unavailable, provider_unavailable, invalid_input, degraded_evidence]
            statusMessage: { type: string }
            retryAt: { type: string }
            completedAt: { type: string }
            organizationId: { type: string }
            scope:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string }
            reused: { type: boolean }
            createdAt: { type: string }
            updatedAt: { type: string }
    ListPagination:
      type: object
      required: [limit, offset, hasMore]
      properties:
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
    WorkflowExecutionListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowExecutionProjection'
        pagination: { $ref: '#/components/schemas/ListPagination' }
...
    WorkflowExecutionProjection:
      type: object
      required: [workflowId, workflowType, namespace, taskQueue, status, organizationId, createdAt, updatedAt]
      properties:
        workflowId: { type: string }
        runId: { type: string }
        workflowType: { type: string }
        blueprintId: { type: string }
        blueprintVersion: { type: string }
        parentWorkflowId: { type: string }
        trigger: { type: string }
        namespace: { type: string }
        taskQueue: { type: string }
        status: { type: string }
        statusReason: { type: string }
        statusMessage: { type: string }
        retryAt: { type: string }
        completedAt: { type: string }
        organizationId: { type: string }
        scope:
          type: object
          required: [ids]
          properties:
            ids:
              type: array
              items: { type: string }
        reused: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    WorkflowStartIntent:
      type: object
      additionalProperties: false
      required: [key]
      properties:
        key: { type: string, minLength: 1, maxLength: 128 }
        businessInput: { type: object }
    WorkflowChangePlan:
      type: object
      additionalProperties: false
      required: [contractVersion, planId, coordinatorId, organizationId, observedAt, changes]
      properties:
        contractVersion:
          const: workflow-change-plan.v1
        planId: { type: string }
        coordinatorId: { type: string }
        organizationId: { type: string }
        projectId: { type: string }
        scope: { $ref: '#/components/schemas/ExecutionScope' }
        observedAt: { type: string }
        metadata: { $ref: '#/components/schemas/WorkflowPlanMetadata' }
        evidenceRefs:
          type: array
          items: { type: string }
        changes:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [kind, reason, requiresApproval]
            properties:
              kind: { enum: [create, update, deprecate, restore, set_current, cancel] }
              targetBlueprintId: { type: string }
              targetBlueprintVersion: { type: string }
              targetWorkflowId: { type: string }
              blueprint: { type: object }
              start: { $ref: '#/components/schemas/WorkflowStartIntent' }
              reason: { type: string }
              evidenceRefs:
                type: array
                items: { type: string }
              requiresApproval: { type: boolean }
    WorkflowPlanMetadata:
      type: object
      additionalProperties: false
      required: [planner]
      properties:
        planner:
          type: object
          additionalProperties: false
          required: [name, version]
          properties:
            name: { type: string, minLength: 1, maxLength: 120 }
            version: { type: string, minLength: 1, maxLength: 80 }
        sourceSchemaVersion: { type: string, minLength: 1, maxLength: 120 }
        promptVersion: { type: string, minLength: 1, maxLength: 120 }
        promptHash: { type: string, pattern: '^[a-f0-9]{64}$' }
    WorkflowChangePlanRequest:
      $ref: '#/components/schemas/WorkflowChangePlan'
    WorkflowPlanValidationResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [planId, organizationId, changeCount, approvalRequired, status, applyStatus]
          properties:
            planId: { type: string }
            organizationId: { type: string }
            changeCount: { type: integer }
            approvalRequired: { type: boolean }
            status: { const: validated_not_applied }
            applyStatus: { const: deferred_persistence_and_approval }
    WorkflowPlanResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/WorkflowPlanRecord' }
    WorkflowPlanListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/WorkflowPlanRecord' }
        pagination: { $ref: '#/components/schemas/ListPagination' }
    WorkflowPlanRecord:
      type: object
      required: [planId, organizationId, coordinatorId, status, approvalRequired, plan, createdAt, updatedAt]
      properties:
        planId: { type: string }
        organizationId: { type: string }
        coordinatorId: { type: string }
        projectId: { type: string }
        status: { enum: [proposed, approved, rejected, applied, expired] }
        approvalRequired: { type: boolean }
        plan: { $ref: '#/components/schemas/WorkflowChangePlanRequest' }
        submittedByUserId: { type: string }
        approvedByUserId: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        approvedAt: { type: string }
        appliedAt: { type: string }